bird-wan-prefix-quirk: Add bird-wan-prefix-quirk role
The role bird-wan-prefix-quirk adds a dummy interface with a /30 ipv4
net containing the wan gateway. This is important when the wan gateway
lies outside the subnet used on the wan interface and routing via the
wan gateway is set up using bird.

Bird checks only the prefixes on interfaces to determine if a nexthop
is reachable. Thus any nexthop that lies outisde all prefix on all
interfaces is considered unreachable even if a /32 p2p route points
to it.