wireguard_p2p: Implement optional glob preference
Since wireguard might be used as a transport by other tunnel
protocols it should be brought up before those. To ensure it
is in fact brought up in the right order the new parameter
ifpref has been added. It acts as a numeric prefix prepended
in front of the wireguard profile name in /etc/network/interfaces.d